Move “Vote” to new component. Move all styles out of PlayGame.module.css and move into component-specific css module files.
