+++ frontend/src/components/GameInvite.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import styles from '../containers/PlayGame.module.css';
+import styles from './GameInvite.module.css';
 
 export function GameInvite({ relativeGameInviteUrl, gameInviteUrl }) {
   return (
