+++ frontend/src/components/GameInvite.module.css
@@ -1,0 +1,4 @@
+.inviteLink a {
+  font-size: 1.5rem;
+  display: inline-block;
+}
