+++ frontend/src/components/GameStory.module.css
@@ -1,0 +1,3 @@
+.story {
+  padding-bottom: 2rem;
+}
