--- frontend/src/components/GameView.jsx
@@ -1,5 +1,0 @@
-import React, { useState, useEffect, useRef } from 'react';
-
-export function GameView() {
-  return <div>GameView</div>;
-}
