+++ frontend/src/components/GameVote.module.css
@@ -11,12 +11,3 @@
 .vote button {
   width: 50px;
 }
-
-.inviteLink a {
-  font-size: 1.5rem;
-  display: inline-block;
-}
-
-.story {
-  padding-bottom: 2rem;
-}
